// typings/machines/google/tasks/sync-list.d.ts
export interface IBind extends IBindBase {
  (
    event: 'Enabled_enter',
    listener: () => boolean | undefined,
    context?: Object
  ): this
  (event: 'Enabled_state', listener: () => any, context?: Object): this
}
/** machine.emit('Enabled', param1, param2) */
export interface IEmit extends IEmitBase {
  (event: 'Enabled_enter'): boolean | void
  (event: 'Enabled_state'): boolean | void
}
// ...
export declare type TStates =
  | 'Enabled'
  | 'Initializing'
  | 'Ready'
  | 'ConfigSet'
  | 'SubsReady'
  | 'SubsInited'
  | 'Reading'
  | 'ReadingDone'
  | 'Cached'
  | 'Dirty'
  | 'QuotaExceeded'
// ...
export interface IState extends IStateBase<TStates> {}
/** Subclassable typesafe state interface */
export interface IStateExt<T extends string> extends IStateBase<T | TStates> {}
export interface IBind extends IBindBase {
  (event: TTransitions, listener: () => boolean | void, context?: Object): this
}
export interface IEmit extends IEmitBase {
  (event: TTransitions): boolean | void
}
export interface IJSONStates {
  Enabled: IState
  Initializing: IState
  Ready: IState
  ConfigSet: IState
  SubsReady: IState
  SubsInited: IState
  Reading: IState
  ReadingDone: IState
  Cached: IState
  Dirty: IState
  QuotaExceeded: IState
  Exception?: IState
}